Add step 4 led code
This code defines several new classes, such as
different types of transitions (immediate, delayed and
sensor-controlled ones), as well as sensors and soon fsm.

This version of the code allows to use the button in order
to immediately switch the led (with interrupting functions).

Note that the file WInterrupts.c is needed to use these functions.

Props to Lucas Escot for the transitions classes and
many interesting discussions.