Begin code for Fsm union
This code aims at building the union of several state machines
used to define behaviors on the circuit. For now, the transitions
are not yet taken into account so this only defines the product
states. Maybe this approach is not the best as it multiplies the
number of states very quickly. This can be tackled by reducing the
number of states but by adding FLAGS and invisible states so that
for example the 7seg has only three states : incrPrint, reset, turnedOff.
Another approach would be to work
with a timed while loop and multiple flags.

There are also some generalizations of the previous classes:
for example, the Actions now take a list of pins and values
to send them at once.